Rename /rate as /rating
This PR renames `/recipes/{id}/rate` as `/recipes/{id}/rating` because URLs describe resources so nouns should be used not verbs. The vebrs are the HTTP methods used to CRUD that resource. This PR also changes the method to `POST` because we want to collect ratings from users, not update a rating field of a recipe.